+++ src/main/java/nokia/internship/packettracer/repo/UserRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface UserRepo extends JpaRepository<User, Long> {
-    Optional<User> findUserById(Long id);
+    Optional<User> findUserById(Integer id);
 }
